Fixed faulty offset in SpriteManager class and fixed issue with component name initialization.

+++ src/BabylonCpp/include/babylon/sprites/sprite_manager.h
@@ -107,7 +107,7 @@ protected:
 
 private:
   void _appendSpriteVertex(size_t index, const Sprite& sprite, int offsetX,
-                           int offsetY, float rowSize);
+                           int offsetY, int rowSize);
 
 public:
   /**
